Misc. scheduler performance improvements
Summary:
Added benchmark for concurrent torrent downloads from multiple peers, and profiled
cpu usage / mutex contention in scheduler, making the following changes:

- Removed unnecessary FileEntry locking on Close.
- Removed `log.Debug` statements from hot code paths (this actually acquires a lock, even if debug is disabled).
- Add on-disk piece completion tracking to make torrent initialization cheaper (it happens on every incoming handshake).

Before:

```
goos: darwin
goarch: amd64
pkg: code.uber.internal/infra/kraken/lib/torrent/scheduler
BenchmarkPieceUploadingAndDownloading-8                1        51544198352 ns/op
PASS
ok      code.uber.internal/infra/kraken/lib/torrent/scheduler   70.735s
```

After:

```
goos: darwin
goarch: amd64
pkg: code.uber.internal/infra/kraken/lib/torrent/scheduler
BenchmarkPieceUploadingAndDownloading-8                1        35272844693 ns/op
PASS
ok      code.uber.internal/infra/kraken/lib/torrent/scheduler   53.670s
```

Reviewers: #kraken, yiran

Reviewed By: #kraken, yiran

Subscribers: yiran, jenkins

Differential Revision: https://code.uberinternal.com/D1326475

+++ torlib/fixtures.go
@@ -65,7 +65,7 @@ type TestTorrentFile struct {
 // CustomTestTorrentFileFixture returns a randomly generated TestTorrentFile
 // of the given size and piece length.
 // TODO(codyg): Move this to storage package.
-func CustomTestTorrentFileFixture(size int, pieceLength int) *TestTorrentFile {
+func CustomTestTorrentFileFixture(size uint64, pieceLength uint64) *TestTorrentFile {
 	f, err := ioutil.TempFile(fixtureTempDir, "torrent_")
 	if err != nil {
 		panic(err)
