+++ utils/randutil/randutil.go
@@ -10,7 +10,7 @@ func init() {
 	rand.Seed(time.Now().UnixNano())
 }
 
-func choose(n int, choices string) []byte {
+func choose(n uint64, choices string) []byte {
 	b := make([]byte, n)
 	for i := range b {
 		c := choices[rand.Intn(len(choices))]
@@ -22,14 +22,14 @@ func choose(n int, choices string) []byte {
 const text = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
 
 // Text returns randomly generated alphanumeric text of length n.
-func Text(n int) []byte {
+func Text(n uint64) []byte {
 	return choose(n, text)
 }
 
 const hex = "0123456789abcdef"
 
 // Hex returns randomly generated hexadecimal string of length n.
-func Hex(n int) string {
+func Hex(n uint64) string {
 	return string(choose(n, hex))
 }
